requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
django-haystack
python-dotenv
Whoosh